+++ Lab2/README.md
@@ -1,0 +1,34 @@
+
+Lab 2
+________________
+
+Individual Lab
+-------------------
+For this lab, we used the openweathermap API to display weather information about a user's specific location. I used HTML geolocation 
+to pull the device's location via coordinates. Then I plugged the saved coordinates into a getJSON function that provided all the necessary
+info needed to call return the JSON. From there, I plugged the JSON components into methods that appended the data to HTML content.
+
+For CSS, I wanted to keep the design simple. Since my theme was 'weather on the go', I wanted the design to be simple enough for a user to get the information 
+they need quickly (No bells and whistles). I originally planned to use an iphone background and display the info within that iphone background;
+however, I thought I could create my own iphone screen border using Bootstrap and ended up doing something completely different. I decided against using 
+animations for the main content since the  weather data was being fetched on page load. If the data takes a little longer than usual to get,
+the animations would be out of synch with the time it takes to display the data.
+
+I am glad we had this lab. Having very little experience with APIs and not doing too well with reading JSON data and outputting them to 
+my HTML pages, this lab helped me get a better understanding of how to use APIs and how to read JSON files. I also feel more a little 
+more comfortable using Javascript now.
+
+
+
+**3 APIs examined:**
+__________________
+    ->API Football
+    (An API that contains data on various soccer information (Players, game stats, team stats, leaderboards). The API can be accessed with a variety 
+    of languages. Requires an account to use but is free for basic API calls.)
+    
+    ->Yahoo Finance
+    (An API that contains a wide variety of financial data. The API is free and can be accessed using a plethora of languages.)
+    
+    ->Coinbase API
+    (An API that contains data on the cryptocurrency market. Information about wallets and transactions can also corraled with this API. A coinbase account 
+    is required to receive an API key and there is a tedious precious one has to go through before being able to actually use the API.)
