+++ Lab3/output.json
@@ -1,0 +1,1 @@
+2620:0:2820:2220:64cb:39d4:9d5e:87e8
