+++ Lab3/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "frempl_quiz1_websci",
+  "name": "shodanapp",
   "version": "1.0.0",
   "description": "",
   "main": "index.js",
@@ -9,7 +9,9 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "express": "^4.17.3",
+    "body-parser": "^1.19.1",
+    "ejs": "^3.1.6",
+    "express": "^4.17.2",
     "node-fetch": "^2.6.6"
   }
 }
