+++ Lab3/views/pages/index.ejs
@@ -1,0 +1,32 @@
+<!DOCTYPE html>
+<html lang="en">
+    <head>
+        <meta charset="utf-8">
+        <meta name="viewport" content="width=device-width, initial-scale=1">
+        <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-1BmE4kWBq78iYhFldvKuhfTAU6auU8tT94WrHftjDbrCEXSU1oBoqyl2QvZ6jIW3" crossorigin="anonymous">
+        <link href="/public/style.css" rel="stylesheet">
+        <script src="https://code.jquery.com/jquery-3.6.0.min.js" integrity="sha256-/xUj+3OJU5yExlq6GSYGSHk7tPXikynS7ogEvDej/m4=" crossorigin="anonymous"></script>
+        <script src="/public/script.js"></script>
+        <script>
+            $(document).ready(function () {
+                $("#submit").click(function () {
+                    $.post("/myapi",
+                {
+                     name: "viSion",
+                     designation: "Professional gamer"
+                },
+                    function (data, status) {
+                        console.log(data);
+                    });
+                });
+            });
+        </script>
+    </head>
+    <body>
+        <h1>LANDING PAGE</h1>
+        <form action="/" method="post">
+            <input type="text" name="ip" placeholder="IP HERE">
+            <button id="submit">submit</button>
+        </form>
+    </body>
+</html>
