+++ README.md
@@ -2,5 +2,3 @@
 
 _______
 **This is my individual repository for Web Science Systems Development**
-
-The quiz was okay. Unfortunately I made a mistake that took me way longer than expected to fix.
