--- index.html
@@ -1,90 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-    <head>
-        <meta charset="utf-8">
-        <meta name="viewport" content="width=device-width, initial-scale=1">
-        <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-1BmE4kWBq78iYhFldvKuhfTAU6auU8tT94WrHftjDbrCEXSU1oBoqyl2QvZ6jIW3" crossorigin="anonymous">
-        <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/js/bootstrap.bundle.min.js" integrity="sha384-ka7Sk0Gln4gmtz2MlQnikT1wXgYsOg+OMhuP+IlRH9sENBO0LRn5q+8nbTov4+1p" crossorigin="anonymous"></script>
-        <script src="https://code.jquery.com/jquery-3.6.0.min.js" integrity="sha256-/xUj+3OJU5yExlq6GSYGSHk7tPXikynS7ogEvDej/m4=" crossorigin="anonymous"></script>
-        <style>
-            #wind_result {
-                color: darkblue;
-                font-size: large;
-            }
-
-            #wind_res_deg {
-                color: darkgreen;
-                font-size: large;
-            }
-
-            #wind_res_speed {
-                color: indianred;
-                font-size: large;
-            }
-        </style>
-        <script>
-            $(document).ready(function () {
-                $("#rpi_wind").click(function () {
-                $.get("/wind",
-                    function (data, status) {
-                        console.log(data);
-                        document.getElementById("wind_result").innerHTML = data.wind.gust;
-                        document.getElementById("wind_res_deg").innerHTML = data.wind.deg;
-                        document.getElementById("wind_res_speed").innerHTML = data.wind.speed;
-                    });
-                });
-            });
-
-            function getWeather(event) {
-                event.preventDefault()
-                var test = document.getElementById('zip').value;
-                $.post("/temperature", 
-                        {
-                            "zip_code": test
-                        },
-                        function (data, status) {
-                            console.log(document.getElementById('zip').value)
-                            console.log(data);
-                            document.getElementById("weatherInfo").innerHTML = data.temp;
-                            if (data.temp < 33) {
-                                document.getElementById("sentence").innerHTML = "IT IS FREEZING";
-                            } else if (data.temp < 50) {
-                                document.getElementById("sentence").innerHTML = "IT IS COLD";
-                            }else if (data.temp < 80) {
-                                document.getElementById("sentence").innerHTML = "IT IS WARM";
-                            } else {
-                                document.getElementById("sentence").innerHTML = "IT IS HOT";
-                            }
-                        })
-                return false
-            }
-                
-        </script>
-    </head>
-    <body>
-        <h1>WEATHER</h1>
-        <button type="button" id="rpi_wind">IS RPI WINDY?</button>
-        <h2>Gust</h2>
-        <div id="wind_result"></div>
-        <h2>Degrees</h2>
-        <div id="wind_res_deg"></div>
-        <h2>Speed</h2>
-        <div id="wind_res_speed"></div>
-
-
-        <form class="m-3" id="zipWeather" onsubmit="return getWeather(event)">
-            <div class="input-group input-group-lg">
-                <div class="input-group-prepend input-group-lg">
-                    <span class="input-group-text rounded-0">GET THE WEATHER</span>
-                </div>
-                <input type="text" name="zip" id="zip" class="form-control rounded-end" placeholder="Enter ZIP code here">
-                <div class="input-group-append input-group-lg ps-2">
-                    <button class="btn rounded-end search" type="submit">SEARCH</button>
-                </div>
-            </div>
-        </form>
-        <h2>Temperature</h2>
-        <div id="weatherInfo"></div>
-        <div id="sentence"></div>
-    </body>
-</html>
